Adds initial auth, teams, bunch of stuff

+++ src/TarokScoreBoard.Api/Controllers/ScoreBoardController.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using TarokScoreBoard.Api.Filters;
+using TarokScoreBoard.Core;
 using TarokScoreBoard.Core.Entities;
 using TarokScoreBoard.Infrastructure.Services;
 using TarokScoreBoard.Shared.DTO;
@@ -13,10 +14,14 @@ namespace TarokScoreBoard.Api.Controllers
   public class ScoreBoardController : BaseController
   {
     private readonly ScoreBoardService scoreboardService;
+    private readonly GameService gameService;
+    private readonly RequestContext context;
 
-    public ScoreBoardController(ScoreBoardService scoreboardService)
+    public ScoreBoardController(ScoreBoardService scoreboardService, GameService gameService, RequestContext context)
     {
       this.scoreboardService = scoreboardService;
+      this.gameService = gameService;
+      this.context = context;
     }
 
     [HttpGet("{gameId}")]
@@ -27,19 +32,54 @@ namespace TarokScoreBoard.Api.Controllers
     }
 
     [HttpPost]
+    [Authorize]
     [TransactionFilter]
     public async Task<ActionResult<ResponseDTO<Round>>> PostRound(CreateRoundDTO createRoundRequest)
     {
+      if (!await CheckTeamId(createRoundRequest.GameId))
+        return Forbid();
+
       var score = await scoreboardService.AddRound(createRoundRequest);
       return Ok(score);
     }
 
+
+
     [HttpDelete("{gameId}")]
+    [Authorize]
     [TransactionFilter]
     public async Task<IActionResult> Delete(Guid gameId)
     {
+      if (!await CheckTeamId(gameId))
+        return Forbid();
+
       var result = await scoreboardService.DeleteLastRound(gameId);
       return NoContent();
+    }
+
+
+    [HttpPost("end/{gameId}")]
+    [Authorize]
+    [TransactionFilter]
+    public async Task<ActionResult<ResponseDTO<Round>>> FinishGame(Guid gameId)
+    {
+      if (!await CheckTeamId(gameId))
+        return Forbid();
+
+      var round = await scoreboardService.EndGame(gameId);
+      return Ok(round);
+    }
+
+
+    [NonAction]
+    private async Task<bool> CheckTeamId(Guid gameId)
+    {
+      var game = await gameService.GetAsync(gameId);
+
+      if (game.TeamId != this.context.TeamId)
+        return false;
+
+      return true;
     }
   }
 }
